backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.20
pillow==11.1.0
# Keep numpy < 2 for widest local compatibility (TensorFlow / older libs).
numpy==1.26.4
joblib>=1.2.0
scikit-learn>=1.6.1,<1.9.0
httpx==0.27.2
pandas>=2.0.0

# ML runtime (needed for soil image classifier)
torch
torchvision


